feat!: refactor enums and export types (#1392)
`ZipFunctionOptions`, `ZipFunctionsOption`, `Config`, and `FunctionConfig` are now exported

BREAKING CHANGE: The exported typescript types for `NodeBundlerType`, `RuntimeType`, and
`ModuleFormat` are not enums anymore but unions of strings. Use the newly exported
`NODE_BUNDLER`, `RUNTIME`, and `MODULE_FORMAT` for constants. Also the types
`NodeBundlerType` and `RuntimeType` were renamed to `NodeBundlerName` and `RuntimeName`
respectively.